refactor: write_message -> outgoing_response_write_message
Besides the name change to reflect the fact that can only be used in
outgoing response mode, the arguments were also refactored and
message.start_line is no longer used.

This change is the first in a series to eliminate the need to parse
start_line and make consistent use of the message concept by no longer
requiring "polymorphic" handling. All content-sensitive data is passed
when the context is in use.