remove close_socket
Clean shutdown (or unclean, if the user desires) is user responsibility
and the library no longer does it.

This commit removes special support for SSL sockets (it no longer needs
the *special* treatment) and open the doors for new types of stream
sockets.

There was the need to specialize close behaviour and this is how it is
being achieved.