remove channel_type
This abstraction was actually pretty useless. The channel_type can be
selected upon the first use from the socket and is not necessary to add
an extra argument to its constructor.

If the need to define channel_type arise, then it'll be implemented
again, but for now it's being removed. The reason for removal is not
lack of usefullness alone, but also that current definition is very
poor.