removed sync versions of all operations.
They weren't implemented, but the reason for continuing to not
implementing them comes from the fact they were designed at a time
when the asynchronous model didn't exist. With the asynchronous model
available, the user can freely use a block completion token to write
all algorithms synchronously, like documented in the N4045 paper.