add basic_buffered_socket whose read buffer is internal
The buffer size still can be customized, changing the template
parameter `N` and it defaults to BOOST_HTTP_SOCKET_DEFAULT_BUFFER_SIZE.

A socket with an internal read buffer is useful for convenience and to
remove a potential source of error (e.g. the user could pass the same
buffer to several sockets).

Changed examples to make use of this new socket.