add test to ensure no reply can be issued while the request isn't ready
A new CompletionToken that adapts asio::yield_context to some future had
to be created. This CompletionToken is use_yielded_future_t and can be
used in other tests within the test/socket.cpp source.

The test purpose is to ensure behaviour created to handle pipelining
issues. Only one <request,response> pair should be handled at a time.
basic_socket's initial write state is empty because this object could
also be used to implement an http client in the future, but once the
server behaviour is assumed, there is no reason to while write requests
for requests that weren't read yet. 100-continue, for instance, can only
be detected after the request has been read.