Add param to BOOST_HTTP_DETAIL_UNREACHABLE to improve readability
A string literal param was added to BOOST_HTTP_DETAIL_UNREACHABLE to
allow custom messages to be printed in case the assertion triggers an
erro.

This commit also updates BOOST_HTTP_DETAIL_UNREACHABLE usage and
improves readability on a few places by giving message strings that
serve as documentation for the implementation.