Refactor code to require an explicit call to `next()` after `set_buffer`
This commit cleans up `set_buffer` behaviour.

The reason to not call `next()` implicitly is to allow the user to
retain current token on buffer for later use when doing some buffer
"compression" (e.g. `realloc`).

This change probably introduces a fix for an untested feature as well
(adding missing `idx = 0`).