Refactor code to always emit token delimiters
Delimiter tokens (end_of_headers, end_of_body and end_of_message) are
part of the program logic and should be emitted every time (assuming no
error happened) to help simplify user code. A small part of this
parser's implementation was simplified as well as a consequence of this
change, but overall more code was required (but not much).

Tests were updated as well.