Add `syntax::chunk_size<CharT>`
This comit introduces the first mini-parser: `chunk_size`. It also
refactors `request_reader` to make use of this new mini-parser.

The mini-parser is stateless. Therefore, now `request_reader` will
reparse characters when it is interrupted in the middle of the chunk
size token.

In general, reading data from the network takes more time than parsing
it. If the user reads large buffers from the network, then re-parsing
will only happen once in a while, and its total performance impact will
be negligible. If the user reads small buffers from the network, then
that will be a performance killer regardless of whether we re-parse or
not.