Convert documentation from Boost's QuickBook to asciidoc
QuickBook is a powerful language, but its complicated tools were
demotivating to create new documentation for Boost.Http. And there was
no support for ePUB generation.

In the search for a new language, I've choose asciidoc as it was
designed to support the semantics of DocBook (like QuickBook) which
allowed a more direct translation from one doc format to another.

Also, the language is nice.