Fix invalid reference on chunkext-related code
If you capture a pointer by-ref in a lambda expression, the pointer to
the pointer is the data-member that the closure object will store. So,
if you try to use the pointer value, another level of indirection goes
through. If the original pointer (not the pointed-to-object) dies, UB
and invalid accesses await.